+++ app/src/main/java/com/example/pam_project/lists/tasks/CreateTaskActivity.java
@@ -29,7 +29,7 @@ public class CreateTaskActivity extends AppCompatActivity {
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-        int id = item.getItemId();
+        long id = item.getItemId();
         final EditText taskTitleInput = findViewById(R.id.create_task_title_input);
         final EditText taskDescriptionInput = findViewById(R.id.create_task_description_input);
